seven-barber/requirements.txt
streamlit
streamlit-option-menu
google-api-python-client
google-auth
google-auth-oauthlib
